Extract method visitor to RememberNavControllerMethodVisitor class
